petBot/requirements.txt
anyio==4.3.0
certifi==2024.2.2
h11==0.14.0
httpcore==1.0.5
idna==3.7
loguru==0.7.2
python-dotenv==1.0.1
sniffio==1.3.1
requests
redis==5.0.8
pycryptodome
nest_asyncio
python-telegram-bot==22.0
APScheduler
python-telegram-bot[job-queue]